File List Processing
- Improve error context
- Remove validators
