+++ submission/submitter/src/main/kotlin/ac/uk/ebi/biostd/submission/processors/SubmissionProjectProcessor.kt
@@ -1,0 +1,14 @@
+package ac.uk.ebi.biostd.submission.processors
+
+import ac.uk.ebi.biostd.submission.exceptions.InvalidProjectException
+import ebi.ac.uk.model.ExtendedSubmission
+import ebi.ac.uk.model.extensions.attachTo
+import ebi.ac.uk.persistence.PersistenceContext
+
+class SubmissionProjectProcessor : SubmissionProcessor {
+    override fun process(submission: ExtendedSubmission, context: PersistenceContext) {
+        submission.attachTo?.let {
+            context.getSubmission(it) ?: throw InvalidProjectException(it)
+        }
+    }
+}
