+++ submission/submitter/src/main/kotlin/ac/uk/ebi/biostd/submission/submitter/SubmissionSubmitter.kt
@@ -1,21 +1,16 @@
 package ac.uk.ebi.biostd.submission.submitter
 
-import ac.uk.ebi.biostd.submission.exceptions.InvalidSubmissionException
+import ac.uk.ebi.biostd.submission.exceptions.SubmissionErrorsContext
 import ac.uk.ebi.biostd.submission.handlers.FilesHandler
 import ac.uk.ebi.biostd.submission.processors.SubmissionProcessor
-import ac.uk.ebi.biostd.submission.validators.SubmissionValidator
-import arrow.core.Try
-import arrow.core.getOrElse
 import ebi.ac.uk.io.sources.FilesSource
 import ebi.ac.uk.model.ExtendedSubmission
 import ebi.ac.uk.model.Submission
 import ebi.ac.uk.persistence.PersistenceContext
-import ebi.ac.uk.util.collections.ifNotEmpty
 import org.springframework.transaction.annotation.Isolation
 import org.springframework.transaction.annotation.Transactional
 
 open class SubmissionSubmitter(
-    private val validators: List<SubmissionValidator>,
     private val processors: List<SubmissionProcessor>,
     private val filesHandler: FilesHandler
 ) {
@@ -23,19 +18,16 @@ open class SubmissionSubmitter(
     open fun submit(
         submission: ExtendedSubmission,
         files: FilesSource,
-        context: PersistenceContext
+        persistenceContext: PersistenceContext
     ): Submission {
-        val exceptionList = mutableListOf<Throwable>()
+        val errorContext = SubmissionErrorsContext()
 
-        validators.map { Try { it.validate(submission, context) }.getOrElse { exceptionList.add(it) } }
-        processors.map { Try { it.process(submission, context) }.getOrElse { exceptionList.add(it) } }
-        Try { filesHandler.processFiles(submission, files) }.getOrElse { exceptionList.add(it) }
+        processors.map { errorContext.runCatching { it.process(submission, persistenceContext) } }
+        errorContext.runCatching { filesHandler.processFiles(submission, files) }
+        errorContext.handleErrors()
 
-        exceptionList.ifNotEmpty {
-            throw InvalidSubmissionException("Submission validation errors", exceptionList)
-        }
+        persistenceContext.deleteSubmissionDrafts(submission)
 
-        context.deleteSubmissionDrafts(submission)
-        return context.saveSubmission(submission)
+        return persistenceContext.saveSubmission(submission)
     }
 }
