--- submission/submitter/src/main/kotlin/ac/uk/ebi/biostd/submission/validators/SubmissionValidator.kt
@@ -1,8 +1,0 @@
-package ac.uk.ebi.biostd.submission.validators
-
-import ebi.ac.uk.model.ExtendedSubmission
-import ebi.ac.uk.persistence.PersistenceContext
-
-interface SubmissionValidator {
-    fun validate(submission: ExtendedSubmission, context: PersistenceContext)
-}
