+++ submission/submitter/src/test/kotlin/ac/uk/ebi/biostd/submission/submitter/SubmissionSubmitterTest.kt
@@ -3,7 +3,6 @@ package ac.uk.ebi.biostd.submission.submitter
 import ac.uk.ebi.biostd.submission.handlers.FilesHandler
 import ac.uk.ebi.biostd.submission.processors.SubmissionProcessor
 import ac.uk.ebi.biostd.submission.test.createBasicExtendedSubmission
-import ac.uk.ebi.biostd.submission.validators.SubmissionValidator
 import ebi.ac.uk.io.sources.FilesSource
 import ebi.ac.uk.persistence.PersistenceContext
 import io.mockk.every
@@ -18,13 +17,12 @@ import org.junit.jupiter.api.extension.ExtendWith
 class SubmissionSubmitterTest(
     @MockK private val filesSource: FilesSource,
     @MockK private val filesHandler: FilesHandler,
-    @MockK private val submissionValidator: SubmissionValidator,
     @MockK private val submissionProcessor: SubmissionProcessor,
     @MockK private val persistenceContext: PersistenceContext
 ) {
     private val submission = createBasicExtendedSubmission()
     private val testInstance =
-        SubmissionSubmitter(listOf(submissionValidator), listOf(submissionProcessor), filesHandler)
+        SubmissionSubmitter(listOf(submissionProcessor), filesHandler)
 
     @BeforeEach
     fun beforeEach() {
@@ -32,7 +30,6 @@ class SubmissionSubmitterTest(
         every { persistenceContext.saveSubmission(submission) } answers { submission }
         every { filesHandler.processFiles(submission, filesSource) } answers { nothing }
         every { submissionProcessor.process(submission, persistenceContext) } answers { nothing }
-        every { submissionValidator.validate(submission, persistenceContext) } answers { nothing }
     }
 
     @Test
@@ -43,7 +40,6 @@ class SubmissionSubmitterTest(
             persistenceContext.saveSubmission(submission)
             filesHandler.processFiles(submission, filesSource)
             submissionProcessor.process(submission, persistenceContext)
-            submissionValidator.validate(submission, persistenceContext)
         }
     }
 }
